Updated download script to only delete global if the download was succesful. Changed default branch to "global". Made file executable.

+++ download-global.sh
@@ -1,10 +1,10 @@
-
+#!/bin/bash
 
 
 #specify branch
 if [[ $# -lt 1 ]]
 then
-	BRANCH='main'
+	BRANCH='global'
 else
 	BRANCH=$1
 fi
@@ -13,12 +13,14 @@ cd /home/pi
 
 #download latest version
 wget --tries=2 https://github.com/The-Redstar/mine-zebro/archive/$BRANCH.zip
-unzip $BRANCH.zip
-rm $BRANCH.zip
+if [[ -f "$BRANCH.zip"]]
+then
+	unzip "$BRANCH.zip"
+	rm "$BRANCH.zip"
 
-rm -r global
-mv mine-zebro-$BRANCH global
-
+	rm -r global
+	mv "mine-zebro-$BRANCH" global
+fi
 
 #create local if it doesn't exist
 mkdir local
